double to int (#1124)
Fix an incorrect type in Geo2RdrParams pybind constructor (double to int).

+++ python/extensions/pybind_isce3/geometry/geo2rdr.cpp
@@ -26,7 +26,7 @@ namespace py = pybind11;
 void addbinding(py::class_<Geo2RdrParams> &pyGeo2RdrParams)
 {
     pyGeo2RdrParams
-        .def(py::init<const double, const double, const double>(),
+        .def(py::init<const double, const int, const double>(),
             py::arg("threshold") = 1e-8,
             py::arg("maxiter") = 50,
             py::arg("delta_range") = 10)
